test: extract progress controller harnesses
