+++ src/test/controllers/ProgressWorkflowActionsController.test.ts
@@ -3,122 +3,23 @@ import { strict as assert } from 'assert';
 
 // Local imports - agent
 import { AgentCategory } from '@agent/core/AgentDataclass';
-import { AgentConfigSchema, type AgentConfig } from '@agent/core/AgentConfig';
-import type { ExecutionRequest } from '@agent/core/executionRequests';
+import type { AgentConfig } from '@agent/core/AgentConfig';
 
 // Local imports - logger
-import type { TaskState, WorkflowTaskState } from '@logger/TaskState';
+import type { TaskState } from '@logger/TaskState';
 
-// Local imports - shared
-import type { OutputFileInfo, StreamTabId } from '@shared/schemas';
-
-// Local imports - controllers
+// Local imports - test support
 import {
-  ProgressWorkflowActionsController,
-  type WorkflowDiffRequest,
-  type WorkflowFileOperation,
-  type WorkflowFileOperationRequest,
-} from '../../controllers/progressView/ProgressWorkflowActionsController';
-
-function createAgentConfig(overrides: Partial<AgentConfig> = {}): AgentConfig {
-  return AgentConfigSchema.parse({
-    agent: 'correct',
-    model: 'gemini31p',
-    inputFile: 'input.tex',
-    outputFiles: ['declared.tex'],
-    agentCategory: AgentCategory.Workflow,
-    ...overrides,
-  });
-}
-
-function createWorkflowTaskState(
-  overrides: Partial<AgentConfig> = {},
-  activeFiles: Partial<WorkflowTaskState['activeFiles']> = {},
-): WorkflowTaskState {
-  return {
-    agentConfig: createAgentConfig({
-      agentCategory: AgentCategory.Workflow,
-      ...overrides,
-    }) as AgentConfig & { agentCategory: AgentCategory.Workflow },
-    activeFiles: {
-      input: true,
-      reference: false,
-      auxiliary: false,
-      media: false,
-      output: true,
-      ...activeFiles,
-    },
-  };
-}
-
-function createOutputFile(
-  absolutePath: string,
-  relativePath = absolutePath,
-): OutputFileInfo {
-  return {
-    source: 'input.tex',
-    location: {
-      kind: 'workspace',
-      absolutePath,
-      relativePath,
-    },
-    round: 1,
-    lineage: null,
-    diff: null,
-  };
-}
-
-function createController(options?: {
-  taskStates?: Map<StreamTabId, TaskState>;
-  executionIds?: Map<StreamTabId, string>;
-  outputs?: Map<StreamTabId, Map<number, OutputFileInfo[]>>;
-  knownWorkspaceOutputs?: Map<StreamTabId, Set<string>>;
-}): {
-  controller: ProgressWorkflowActionsController;
-  executed: ExecutionRequest[];
-  diffs: WorkflowDiffRequest[];
-  fileOperations: Array<{
-    operation: WorkflowFileOperation;
-    request: WorkflowFileOperationRequest;
-  }>;
-} {
-  const executed: ExecutionRequest[] = [];
-  const diffs: WorkflowDiffRequest[] = [];
-  const fileOperations: Array<{
-    operation: WorkflowFileOperation;
-    request: WorkflowFileOperationRequest;
-  }> = [];
-
-  return {
-    controller: new ProgressWorkflowActionsController({
-      state: {
-        getTaskState: (stream) => options?.taskStates?.get(stream),
-        getExecutionId: (stream) => options?.executionIds?.get(stream),
-        getOutputFiles: (stream) =>
-          new Map(options?.outputs?.get(stream) ?? []),
-        getKnownWorkspaceOutputPaths: (stream) =>
-          new Set(options?.knownWorkspaceOutputs?.get(stream) ?? []),
-      },
-      executeAgent: async (request) => {
-        executed.push(request);
-      },
-      runDiff: async (request) => {
-        diffs.push(request);
-      },
-      runFileOperation: async (operation, request) => {
-        fileOperations.push({ operation, request });
-      },
-    }),
-    executed,
-    diffs,
-    fileOperations,
-  };
-}
+  createAgentConfig,
+  createOutputFile,
+  createProgressWorkflowActionsHarness,
+  createWorkflowTaskState,
+} from '../support/ProgressControllerHarnesses';
 
 describe('ProgressWorkflowActionsController', () => {
   it('resumes workflow streams with their execution id', async () => {
     const taskState = createWorkflowTaskState();
-    const { controller, executed } = createController({
+    const { controller, executed } = createProgressWorkflowActionsHarness({
       taskStates: new Map([['stream-a', taskState]]),
       executionIds: new Map([['stream-a', 'exec-123']]),
     });
@@ -132,7 +33,7 @@ describe('ProgressWorkflowActionsController', () => {
 
   it('runs a fresh request without the existing execution id', async () => {
     const taskState = createWorkflowTaskState();
-    const { controller, executed } = createController({
+    const { controller, executed } = createProgressWorkflowActionsHarness({
       taskStates: new Map([['stream-a', taskState]]),
       executionIds: new Map([['stream-a', 'exec-123']]),
     });
@@ -149,9 +50,10 @@ describe('ProgressWorkflowActionsController', () => {
         outputFiles: [],
       }) as AgentConfig & { agentCategory: AgentCategory.ToolUse },
     };
-    const { controller, diffs, fileOperations } = createController({
-      taskStates: new Map([['stream-a', toolUseState]]),
-    });
+    const { controller, diffs, fileOperations } =
+      createProgressWorkflowActionsHarness({
+        taskStates: new Map([['stream-a', toolUseState]]),
+      });
 
     await controller.diffStream('stream-a');
     await controller.runFileOperation('stream-a', 'pack');
@@ -169,7 +71,7 @@ describe('ProgressWorkflowActionsController', () => {
       { outputFiles: ['declared.tex'] },
       { output: false },
     );
-    const { controller, diffs } = createController({
+    const { controller, diffs } = createProgressWorkflowActionsHarness({
       taskStates: new Map([['stream-a', taskState]]),
       executionIds: new Map([['stream-a', 'exec-123']]),
       outputs: new Map([['stream-a', new Map([[1, [output]]])]]),
@@ -197,13 +99,15 @@ describe('ProgressWorkflowActionsController', () => {
       outputFiles: ['declared.tex', '/workspace/generated.tex'],
       useMultipleOutputs: true,
     });
-    const { controller, fileOperations } = createController({
-      taskStates: new Map([['stream-a', taskState]]),
-      executionIds: new Map([['stream-a', 'exec-123']]),
-      knownWorkspaceOutputs: new Map([
-        ['stream-a', new Set(['/workspace/generated.tex', 'extra.tex'])],
-      ]),
-    });
+    const { controller, fileOperations } = createProgressWorkflowActionsHarness(
+      {
+        taskStates: new Map([['stream-a', taskState]]),
+        executionIds: new Map([['stream-a', 'exec-123']]),
+        knownWorkspaceOutputs: new Map([
+          ['stream-a', new Set(['/workspace/generated.tex', 'extra.tex'])],
+        ]),
+      },
+    );
 
     await controller.runFileOperation('stream-a', 'pack');
 
@@ -236,12 +140,14 @@ describe('ProgressWorkflowActionsController', () => {
       },
       { output: true },
     );
-    const { controller, fileOperations } = createController({
-      taskStates: new Map([['stream-a', taskState]]),
-      knownWorkspaceOutputs: new Map([
-        ['stream-a', new Set(['generated.tex'])],
-      ]),
-    });
+    const { controller, fileOperations } = createProgressWorkflowActionsHarness(
+      {
+        taskStates: new Map([['stream-a', taskState]]),
+        knownWorkspaceOutputs: new Map([
+          ['stream-a', new Set(['generated.tex'])],
+        ]),
+      },
+    );
 
     await controller.runFileOperation('stream-a', 'clean');
